pagetable

+++ os161/kern/main/menu.c
@@ -115,7 +115,7 @@ common_prog(int nargs, char **args) {
     struct trapframe* tf = kmalloc(sizeof(struct trapframe));
     int x;
     tf->tf_a0 = childthread->pid;
-    tf->tf_a1 = &x;
+    tf->tf_a1 = (u_int32_t) &x;
     tf->tf_a2 = 0;
     sys_waitpid(tf,1);      // wait for the program to exit
     kfree(tf);
